Implement Extension mechanism throughout core library
