+++ lib/flutter_html.dart
@@ -1,27 +1,25 @@
 library flutter_html;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_html/custom_render.dart';
-import 'package:flutter_html/html_parser.dart';
+import 'package:flutter_html/src/html_parser.dart';
+import 'package:flutter_html/src/extension/extension.dart';
 import 'package:flutter_html/src/html_elements.dart';
-import 'package:flutter_html/style.dart';
+import 'package:flutter_html/src/style.dart';
 import 'package:html/dom.dart' as dom;
 
-export 'package:flutter_html/custom_render.dart';
 //export render context api
-export 'package:flutter_html/html_parser.dart';
-//export render context api
-export 'package:flutter_html/html_parser.dart';
+export 'package:flutter_html/src/html_parser.dart';
 //export src for advanced custom render uses (e.g. casting context.tree)
 export 'package:flutter_html/src/anchor.dart';
-export 'package:flutter_html/src/interactable_element.dart';
-export 'package:flutter_html/src/layout_element.dart';
-export 'package:flutter_html/src/replaced_element.dart';
-export 'package:flutter_html/src/styled_element.dart';
-//export css_box_widget for use in custom render.
+export 'package:flutter_html/src/tree/interactable_element.dart';
+export 'package:flutter_html/src/tree/replaced_element.dart';
+export 'package:flutter_html/src/tree/styled_element.dart';
+//export css_box_widget for use in extensions.
 export 'package:flutter_html/src/css_box_widget.dart';
 //export style api
-export 'package:flutter_html/style.dart';
+export 'package:flutter_html/src/style.dart';
+//export extension api
+export 'package:flutter_html/src/extension/extension.dart';
 
 class Html extends StatefulWidget {
   /// The `Html` widget takes HTML as input and displays a RichText
@@ -33,9 +31,6 @@ class Html extends StatefulWidget {
   ///
   /// **onLinkTap** This function is called whenever a link (`<a href>`)
   /// is tapped.
-  /// **customRender** This function allows you to return your own widgets
-  /// for existing or custom HTML tags.
-  /// See [its wiki page](https://github.com/Sub6Resources/flutter_html/wiki/All-About-customRender) for more info.
   ///
   /// **onImageError** This is called whenever an image fails to load or
   /// display on the page.
@@ -55,7 +50,7 @@ class Html extends StatefulWidget {
     required this.data,
     this.onLinkTap,
     this.onAnchorTap,
-    this.customRenders = const {},
+    this.extensions = const [],
     this.onCssParseError,
     this.onImageError,
     this.shrinkWrap = false,
@@ -73,7 +68,7 @@ class Html extends StatefulWidget {
     @required dom.Document? document,
     this.onLinkTap,
     this.onAnchorTap,
-    this.customRenders = const {},
+    this.extensions = const [],
     this.onCssParseError,
     this.onImageError,
     this.shrinkWrap = false,
@@ -92,7 +87,7 @@ class Html extends StatefulWidget {
     @required this.documentElement,
     this.onLinkTap,
     this.onAnchorTap,
-    this.customRenders = const {},
+    this.extensions = const [],
     this.onCssParseError,
     this.onImageError,
     this.shrinkWrap = false,
@@ -136,9 +131,9 @@ class Html extends StatefulWidget {
   /// A list of HTML tags that are the only tags that are rendered. By default, this list is empty and all supported HTML tags are rendered.
   final List<String> tagsList;
 
-  /// Either return a custom widget for specific node types or return null to
-  /// fallback to the default rendering.
-  final Map<CustomRenderMatcher, CustomRender> customRenders;
+  /// A list of [Extension]s that add additional capabilities to flutter_html
+  /// See the [Extension] class for more details.
+  final List<Extension> extensions;
 
   /// An API that allows you to override the default style for any HTML element
   final Map<String, Style> style;
@@ -147,8 +142,6 @@ class Html extends StatefulWidget {
     ..addAll(HtmlElements.interactableElements)
     ..addAll(HtmlElements.replacedElements)
     ..addAll(HtmlElements.layoutElements)
-    ..addAll(HtmlElements.tableCellElements)
-    ..addAll(HtmlElements.tableDefinitionElements)
     ..addAll(HtmlElements.externalElements);
 
   @override
@@ -188,196 +181,9 @@ class _HtmlState extends State<Html> {
       onCssParseError: widget.onCssParseError,
       onImageError: widget.onImageError,
       shrinkWrap: widget.shrinkWrap,
-      selectable: false,
       style: widget.style,
-      customRenders: {}
-        ..addAll(widget.customRenders)
-        ..addAll(generateDefaultRenders()),
+      extensions: widget.extensions,
       tagsList: widget.tagsList.isEmpty ? Html.tags : widget.tagsList,
-    );
-  }
-}
-
-class SelectableHtml extends StatefulWidget {
-  /// The `SelectableHtml` widget takes HTML as input and displays a RichText
-  /// tree of the parsed HTML content (which is selectable)
-  ///
-  /// **Attributes**
-  /// **data** *required* takes in a String of HTML data (required only for `Html` constructor).
-  /// **documentElement** *required* takes in a Element of HTML data (required only for `Html.fromDom` and `Html.fromElement` constructor).
-  ///
-  /// **onLinkTap** This function is called whenever a link (`<a href>`)
-  /// is tapped.
-  ///
-  /// **onAnchorTap** This function is called whenever an anchor (#anchor-id)
-  /// is tapped.
-  ///
-  /// **tagsList** Tag names in this array will be the only tags rendered. By default, all tags that support selectable content are rendered.
-  ///
-  /// **style** Pass in the style information for the Html here.
-  /// See [its wiki page](https://github.com/Sub6Resources/flutter_html/wiki/Style) for more info.
-  ///
-  /// **PLEASE NOTE**
-  ///
-  /// There are a few caveats due to Flutter [#38474](https://github.com/flutter/flutter/issues/38474):
-  ///
-  /// 1. The list of tags that can be rendered is significantly reduced.
-  /// Key omissions include no support for images/video/audio, table, and ul/ol because they all require widgets and `WidgetSpan`s.
-  ///
-  /// 2. No support for `customRender`, `customImageRender`, `onImageError`, `onImageTap`, `onMathError`, and `navigationDelegateForIframe`.
-  ///
-  /// 3. Styling support is significantly reduced. Only text-related styling works
-  /// (e.g. bold or italic), while container related styling (e.g. borders or padding/margin)
-  /// do not work because we can't use the `ContainerSpan` class (it needs an enclosing `WidgetSpan`).
-
-  SelectableHtml({
-    Key? key,
-    GlobalKey? anchorKey,
-    required this.data,
-    this.onLinkTap,
-    this.onAnchorTap,
-    this.onCssParseError,
-    this.shrinkWrap = false,
-    this.style = const {},
-    this.customRenders = const {},
-    this.tagsList = const [],
-    this.selectionControls,
-    this.scrollPhysics,
-  })  : documentElement = null,
-        assert(data != null),
-        _anchorKey = anchorKey ?? GlobalKey(),
-        super(key: key);
-
-  SelectableHtml.fromDom({
-    Key? key,
-    GlobalKey? anchorKey,
-    @required dom.Document? document,
-    this.onLinkTap,
-    this.onAnchorTap,
-    this.onCssParseError,
-    this.shrinkWrap = false,
-    this.style = const {},
-    this.customRenders = const {},
-    this.tagsList = const [],
-    this.selectionControls,
-    this.scrollPhysics,
-  })  : data = null,
-        assert(document != null),
-        documentElement = document!.documentElement,
-        _anchorKey = anchorKey ?? GlobalKey(),
-        super(key: key);
-
-  SelectableHtml.fromElement({
-    Key? key,
-    GlobalKey? anchorKey,
-    @required this.documentElement,
-    this.onLinkTap,
-    this.onAnchorTap,
-    this.onCssParseError,
-    this.shrinkWrap = false,
-    this.style = const {},
-    this.customRenders = const {},
-    this.tagsList = const [],
-    this.selectionControls,
-    this.scrollPhysics,
-  })  : data = null,
-        assert(documentElement != null),
-        _anchorKey = anchorKey ?? GlobalKey(),
-        super(key: key);
-
-  /// A unique key for this Html widget to ensure uniqueness of anchors
-  final GlobalKey _anchorKey;
-
-  /// The HTML data passed to the widget as a String
-  final String? data;
-
-  /// The HTML data passed to the widget as a pre-processed [dom.Element]
-  final dom.Element? documentElement;
-
-  /// A function that defines what to do when a link is tapped
-  final OnTap? onLinkTap;
-
-  /// A function that defines what to do when an anchor link is tapped. When this value is set,
-  /// the default anchor behaviour is overwritten.
-  final OnTap? onAnchorTap;
-
-  /// A function that defines what to do when CSS fails to parse
-  final OnCssParseError? onCssParseError;
-
-  /// A parameter that should be set when the HTML widget is expected to be
-  /// have a flexible width, that doesn't always fill its maximum width
-  /// constraints. For example, auto horizontal margins are ignored, and
-  /// block-level elements only take up the width they need.
-  final bool shrinkWrap;
-
-  /// A list of HTML tags that are the only tags that are rendered. By default, this list is empty and all supported HTML tags are rendered.
-  final List<String> tagsList;
-
-  /// An API that allows you to override the default style for any HTML element
-  final Map<String, Style> style;
-
-  /// Custom Selection controls allows you to override default toolbar and build custom toolbar
-  /// options
-  final TextSelectionControls? selectionControls;
-
-  /// Allows you to override the default scrollPhysics for [SelectableText.rich]
-  final ScrollPhysics? scrollPhysics;
-
-  /// Either return a custom widget for specific node types or return null to
-  /// fallback to the default rendering.
-  final Map<CustomRenderMatcher, SelectableCustomRender> customRenders;
-
-  static List<String> get tags =>
-      List<String>.from(HtmlElements.selectableElements);
-
-  @override
-  State<StatefulWidget> createState() => _SelectableHtmlState();
-}
-
-class _SelectableHtmlState extends State<SelectableHtml> {
-  late dom.Element documentElement;
-
-  @override
-  void initState() {
-    super.initState();
-    documentElement = widget.data != null
-        ? HtmlParser.parseHTML(widget.data!)
-        : widget.documentElement!;
-  }
-  @override
-  void didUpdateWidget(SelectableHtml oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if ((widget.data != null && oldWidget.data != widget.data) ||
-        oldWidget.documentElement != widget.documentElement) {
-      documentElement = widget.data != null
-          ? HtmlParser.parseHTML(widget.data!)
-          : widget.documentElement!;
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: widget.shrinkWrap ? null : MediaQuery.of(context).size.width,
-      child: HtmlParser(
-        key: widget._anchorKey,
-        htmlData: documentElement,
-        onLinkTap: widget.onLinkTap,
-        onAnchorTap: widget.onAnchorTap,
-        onImageTap: null,
-        onCssParseError: widget.onCssParseError,
-        onImageError: null,
-        shrinkWrap: widget.shrinkWrap,
-        selectable: true,
-        style: widget.style,
-        customRenders: {}
-          ..addAll(widget.customRenders)
-          ..addAll(generateDefaultRenders()),
-        tagsList:
-            widget.tagsList.isEmpty ? SelectableHtml.tags : widget.tagsList,
-        selectionControls: widget.selectionControls,
-        scrollPhysics: widget.scrollPhysics,
-      ),
     );
   }
 }
