+++ lib/src/anchor.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_html/src/styled_element.dart';
+import 'package:flutter_html/src/tree/styled_element.dart';
 
 class AnchorKey extends GlobalKey {
   static final Set<AnchorKey> _registry = <AnchorKey>{};
