+++ lib/src/extension/extension.dart
@@ -2,45 +2,48 @@ import 'package:flutter/painting.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_html/src/extension/extension_context.dart';
 
+export 'package:flutter_html/src/extension/extension_context.dart';
+
 /// The [Extension] class allows you to customize the behavior of flutter_html
 /// or add additional functionality.
 ///
 /// TODO add additional documentation
 ///
 abstract class Extension {
+  const Extension();
+
   /// Tells the [HtmlParser] what additional tags to add to the default
   /// supported tag list (the extension's user can still override this by
   /// setting an explicit tagList on the Html widget).
   ///
   /// Extension creators should override this with any additional tags
   /// that should be visible to the end user.
-  List<String> get supportedTags;
+  Set<String> get supportedTags;
 
   /// This method is called to test whether or not this extension needs to do
   /// any work in this context.
   ///
-  /// Subclasses must override this method and return true if they'd like the
-  /// other methods to be called in a certain context.
-  bool matches(ExtensionContext context);
-
-  // Converts parsed HTML to a StyledElement. Need to define default behavior, or perhaps defer this step back to the Html widget by default
-  StyledElement lex(ExtensionContext context) {
-    throw UnimplementedError("TODO");
+  /// By default returns true if [supportedTags] contains the element's name
+  bool matches(ExtensionContext context) {
+    return supportedTags.contains(context.elementName);
   }
 
-  // Called before styles are applied to the tree. Default behavior: return tree;
-  StyledElement beforeStyle(ExtensionContext context) {
-    return context.styledElement!;
+  // Converts parsed HTML to a StyledElement.
+  StyledElement lex(ExtensionContext context, List<StyledElement> children) {
+    throw UnimplementedError("Extension `$runtimeType` matched `${context.elementName}` but didn't implement `lex`");
   }
 
-  // Called after styling, but before extra elements/whitespace has been removed, margins collapsed, list characters processed, or relative values calculated. Default behavior: return tree;
-  StyledElement beforeProcessing(ExtensionContext context) {
-    return context.styledElement!;
-  }
+  // Called before styles are applied to the tree. Default behavior: do nothing;
+  void beforeStyle(ExtensionContext context) {}
 
-  //The final step in the chain. Converts the StyledElement tree, with its attached `Style` elements, into an `InlineSpan` tree that includes Widget/TextSpans that can be rendered in a RichText or Text.rich widget. Need to define default behavior, or perhaps defer this step back to the Html widget by default
-  InlineSpan parse(ExtensionContext context) {
-    throw UnimplementedError("TODO");
+  // Called after styling, but before extra elements/whitespace has been
+  // removed, margins collapsed, list characters processed, or relative
+  // values calculated. Default behavior: do nothing;
+  void beforeProcessing(ExtensionContext context) {}
+
+  //The final step in the chain. Converts the StyledElement tree, with its attached `Style` elements, into an `InlineSpan` tree that includes Widget/TextSpans that can be rendered in a RichText widget.
+  InlineSpan parse(ExtensionContext context, Map<StyledElement, InlineSpan> Function() parseChildren) {
+    throw UnimplementedError("Extension `$runtimeType` matched `${context.styledElement!.name}` but didn't implement `parse`");
   }
 
   //Called when the Html widget is being destroyed. This would be a very good place to dispose() any controllers or free any resources that the extension uses. Default behavior: do nothing.
