+++ test/html_parser_test.dart
@@ -33,143 +33,6 @@ void main() {
 void testNewParser(BuildContext context) {
   HtmlParser.parseHTML("<b>Hello, World!</b>");
 
-  HtmlParser.lexDomTree(
-    HtmlParser.parseHTML("Hello! <b>Hello, World!</b><i>Hello, New World!</i>"),
-    [],
-    Html.tags,
-    context,
-    HtmlParser(
-      key: null,
-      htmlData: HtmlParser.parseHTML(
-          "Hello! <b>Hello, World!</b><i>Hello, New World!</i>"),
-      onLinkTap: null,
-      onAnchorTap: null,
-      onImageTap: null,
-      onCssParseError: null,
-      onImageError: null,
-      shrinkWrap: false,
-      selectable: true,
-      style: const {},
-      customRenders: generateDefaultRenders(),
-      tagsList: Html.tags,
-      selectionControls: null,
-      scrollPhysics: null,
-    ),
-  );
-
-  HtmlParser.lexDomTree(
-    HtmlParser.parseHTML(
-        "Hello, World! <a href='https://example.com'>This is a link</a>"),
-    [],
-    Html.tags,
-    context,
-    HtmlParser(
-      key: null,
-      htmlData: HtmlParser.parseHTML(
-          "Hello, World! <a href='https://example.com'>This is a link</a>"),
-      onLinkTap: null,
-      onAnchorTap: null,
-      onImageTap: null,
-      onCssParseError: null,
-      onImageError: null,
-      shrinkWrap: false,
-      selectable: true,
-      style: const {},
-      customRenders: generateDefaultRenders(),
-      tagsList: Html.tags,
-      selectionControls: null,
-      scrollPhysics: null,
-    ),
-  );
-
-  HtmlParser.lexDomTree(
-    HtmlParser.parseHTML("<img alt='' src='https://image.example.com' />"),
-    [],
-    Html.tags,
-    context,
-    HtmlParser(
-      key: null,
-      htmlData: HtmlParser.parseHTML(
-          "<img alt='' src='https://image.example.com' />"),
-      onLinkTap: null,
-      onAnchorTap: null,
-      onImageTap: null,
-      onCssParseError: null,
-      onImageError: null,
-      shrinkWrap: false,
-      selectable: true,
-      style: const {},
-      customRenders: generateDefaultRenders(),
-      tagsList: Html.tags,
-      selectionControls: null,
-      scrollPhysics: null,
-    ),
-  );
-
-  HtmlParser.lexDomTree(
-    HtmlParser.parseHTML(
-        "<div><div><div><div><a href='https://example.com'>Link</a><div>Hello, World! <b>Bold and <i>Italic</i></b></div></div></div></div></div>"),
-    [],
-    Html.tags,
-    context,
-    HtmlParser(
-      key: null,
-      htmlData: HtmlParser.parseHTML(
-          "<div><div><div><div><a href='https://example.com'>Link</a><div>Hello, World! <b>Bold and <i>Italic</i></b></div></div></div></div></div>"),
-      onLinkTap: null,
-      onAnchorTap: null,
-      onImageTap: null,
-      onCssParseError: null,
-      onImageError: null,
-      shrinkWrap: false,
-      selectable: true,
-      style: const {},
-      customRenders: generateDefaultRenders(),
-      tagsList: Html.tags,
-      selectionControls: null,
-      scrollPhysics: null,
-    ),
-  );
-
-  /*ReplacedElement videoContentElement = parseReplacedElement(
-    HtmlParser.parseHTML("""
-      <video width="320" height="240" controls>
-       <source src="movie.mp4" type="video/mp4">
-       <source src="movie.ogg" type="video/ogg">
-       Your browser does not support the video tag.
-      </video>
-    """).getElementsByTagName("video")[0],
-    [],
-    null,
-  );
-
-  expect(videoContentElement, isA<VideoContentElement>());
-  if (videoContentElement is VideoContentElement) {
-    expect(videoContentElement.showControls, equals(true),
-        reason: "Controls isn't working");
-    expect(videoContentElement.src, hasLength(2),
-        reason: "Not enough sources...");
-  }
-
-  ReplacedElement audioContentElement = parseReplacedElement(
-    HtmlParser.parseHTML("""
-      <audio controls>
-        <source src='audio.mp3' type='audio/mpeg'>
-        <source src='audio.wav' type='audio/wav'>
-        Your browser does not support the audio tag.
-      </audio>
-    """).getElementsByTagName("audio")[0],
-    [],
-    null,
-  );
-  expect(audioContentElement, isA<AudioContentElement>());
-  if (audioContentElement is AudioContentElement) {
-    expect(audioContentElement.showControls, equals(true),
-        reason: "Controls isn't working");
-    expect(audioContentElement.src, hasLength(2),
-        reason: "Not enough sources...");
-  }*/
-
   Style style1 = Style(
     display: Display.block,
     fontWeight: FontWeight.bold,
